Backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.27.2
pydantic==2.10.3
pyspellchecker==0.8.1
huggingface_hub==0.26.5
Pillow==11.0.0
pytest==7.3.1
pytest-asyncio==0.20.3
pytest-cov==4.0.0
